+++ CHANGELOG.md
@@ -37,6 +37,9 @@ He also sells TM70 Flash.
 * An item to allow use of field moves without a party member knowing them. (sprites missing)
 * An option to just talk to trainers. (useful for trainersanity)
 * Added option groups to the WebWorld and the template YAML ([@snowflav-goob](https://github.com/snowflav-goob))
+* The slot name is now stored in the ROM, so it does not need to be entered when BizHawk Client connects.
+* There is now an option to have "relaxed" player and rival name conversion, which replaces unknown characters and truncates
+appropriately.
 ### Changed
 * In-game game options option is now validated before generation.
 * Some evolution stones have been key-itemified, and are no longer consumed when used.